Renamed package members and its usages
